LangChain/requirements.txt
langchain
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
lxml
langchain-text-splitters
streamlit